requirements.txt
openai
python-dotenv==1.0.1
xmltodict
psutil
